python/requirements.txt
pygame
pyserial
enum